feat: Add POST endpoint for summing numbers
This commit adds a new POST endpoint '/sum1' to the AppController class. The endpoint accepts an array of numbers in the request body and uses axios to make a POST request to 'http://localhost:3001/math/sum' with the numbers. The response data is returned if the request is successful, otherwise the error response data is returned.

Co-authored-by: [Author Name]